Moves tcp errors to errors.cpp
